+++ packages/contracts/contracts/test/plugin-mock/Clonable/PluginCloneableSetupMock.sol
@@ -28,7 +28,7 @@ contract PluginCloneableSetupV1Mock is PluginSetup {
         returns (
             address plugin,
             address[] memory helpers,
-            PermissionLib.ItemMultiTarget[] memory permissions
+            PermissionLib.AdvancedPermission[] memory permissions
         )
     {
         plugin = mockPluginProxy(pluginBase, _dao);
@@ -47,7 +47,7 @@ contract PluginCloneableSetupV1Mock is PluginSetup {
         address _plugin,
         address[] calldata _currentHelpers,
         bytes calldata
-    ) external virtual override returns (PermissionLib.ItemMultiTarget[] memory permissions) {
+    ) external virtual override returns (PermissionLib.AdvancedPermission[] memory permissions) {
         (_dao, _plugin, _currentHelpers);
         permissions = mockPermissions(5, 6, PermissionLib.Operation.Revoke);
     }
@@ -76,7 +76,7 @@ contract PluginCloneableSetupV2Mock is PluginCloneableSetupV1Mock {
         returns (
             address plugin,
             address[] memory helpers,
-            PermissionLib.ItemMultiTarget[] memory permissions
+            PermissionLib.AdvancedPermission[] memory permissions
         )
     {
         plugin = mockPluginProxy(pluginBase, _dao);
@@ -95,7 +95,7 @@ contract PluginCloneableSetupV2Mock is PluginCloneableSetupV1Mock {
         address _plugin,
         address[] calldata _currentHelpers,
         bytes calldata
-    ) external virtual override returns (PermissionLib.ItemMultiTarget[] memory permissions) {
+    ) external virtual override returns (PermissionLib.AdvancedPermission[] memory permissions) {
         (_dao, _plugin, _currentHelpers);
         permissions = mockPermissions(5, 7, PermissionLib.Operation.Revoke);
     }