+++ packages/contracts/contracts/voting/addresslist/AddresslistVotingSetup.sol
@@ -36,7 +36,7 @@ contract AddresslistVotingSetup is PluginSetup {
         returns (
             address plugin,
             address[] memory helpers,
-            PermissionLib.ItemMultiTarget[] memory permissions
+            PermissionLib.AdvancedPermission[] memory permissions
         )
     {
         IDAO dao = IDAO(_dao);
@@ -60,11 +60,11 @@ contract AddresslistVotingSetup is PluginSetup {
         (helpers); // silence the warning.
 
         // Prepare permissions
-        permissions = new PermissionLib.ItemMultiTarget[](4);
+        permissions = new PermissionLib.AdvancedPermission[](4);
 
         // Set permissions to be granted.
         // Grant the list of prmissions of the plugin to the DAO.
-        permissions[0] = PermissionLib.ItemMultiTarget(
+        permissions[0] = PermissionLib.AdvancedPermission(
             PermissionLib.Operation.Grant,
             plugin,
             _dao,
@@ -72,7 +72,7 @@ contract AddresslistVotingSetup is PluginSetup {
             addresslistVotingBase.UPDATE_ADDRESSES_PERMISSION_ID()
         );
 
-        permissions[1] = PermissionLib.ItemMultiTarget(
+        permissions[1] = PermissionLib.AdvancedPermission(
             PermissionLib.Operation.Grant,
             plugin,
             _dao,
@@ -80,7 +80,7 @@ contract AddresslistVotingSetup is PluginSetup {
             addresslistVotingBase.UPDATE_VOTING_SETTINGS_PERMISSION_ID()
         );
 
-        permissions[2] = PermissionLib.ItemMultiTarget(
+        permissions[2] = PermissionLib.AdvancedPermission(
             PermissionLib.Operation.Grant,
             plugin,
             _dao,
@@ -89,7 +89,7 @@ contract AddresslistVotingSetup is PluginSetup {
         );
 
         // Grant `EXECUTE_PERMISSION` of the DAO to the plugin.
-        permissions[3] = PermissionLib.ItemMultiTarget(
+        permissions[3] = PermissionLib.AdvancedPermission(
             PermissionLib.Operation.Grant,
             _dao,
             plugin,
@@ -109,12 +109,12 @@ contract AddresslistVotingSetup is PluginSetup {
         address _plugin,
         address[] calldata,
         bytes calldata
-    ) external view returns (PermissionLib.ItemMultiTarget[] memory permissions) {
+    ) external view returns (PermissionLib.AdvancedPermission[] memory permissions) {
         // Prepare permissions
-        permissions = new PermissionLib.ItemMultiTarget[](4);
+        permissions = new PermissionLib.AdvancedPermission[](4);
 
         // Set permissions to be Revoked.
-        permissions[0] = PermissionLib.ItemMultiTarget(
+        permissions[0] = PermissionLib.AdvancedPermission(
             PermissionLib.Operation.Revoke,
             _plugin,
             _dao,
@@ -122,7 +122,7 @@ contract AddresslistVotingSetup is PluginSetup {
             addresslistVotingBase.UPDATE_ADDRESSES_PERMISSION_ID()
         );
 
-        permissions[1] = PermissionLib.ItemMultiTarget(
+        permissions[1] = PermissionLib.AdvancedPermission(
             PermissionLib.Operation.Revoke,
             _plugin,
             _dao,
@@ -130,7 +130,7 @@ contract AddresslistVotingSetup is PluginSetup {
             addresslistVotingBase.UPDATE_VOTING_SETTINGS_PERMISSION_ID()
         );
 
-        permissions[2] = PermissionLib.ItemMultiTarget(
+        permissions[2] = PermissionLib.AdvancedPermission(
             PermissionLib.Operation.Revoke,
             _plugin,
             _dao,
@@ -138,7 +138,7 @@ contract AddresslistVotingSetup is PluginSetup {
             addresslistVotingBase.UPGRADE_PLUGIN_PERMISSION_ID()
         );
 
-        permissions[3] = PermissionLib.ItemMultiTarget(
+        permissions[3] = PermissionLib.AdvancedPermission(
             PermissionLib.Operation.Revoke,
             _dao,
             _plugin,